Split delegations from mixnodes into separate module
