+++ contracts/mixnet/src/delegations/mod.rs
@@ -1,7 +1,7 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod bonding_queries;
-pub mod layer_queries;
-pub mod storage;
-pub mod transactions;
+pub(crate) mod helpers;
+pub(crate) mod queries;
+pub(crate) mod storage;
+pub(crate) mod transactions;
