+++ contracts/mixnet/src/lib.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod contract;
+mod delegations;
 pub mod error;
 pub mod gateways;
 pub mod mixnet_contract_settings;
